Upgrade (almost) to null safety
I can't quite upgrade to null safety yet due to columnar_ouput depending on a old version of args, but this gets us close.

This also forced upgrading (4?!) versions of package:graphql so the syntax around error handling is slightly different (errors are no longer strings but rather actual Error objects which can be thrown).  Fixed the graphql code to also throw Error objects instead of strings as is modern dart (or at least my understanding of it).

Wrapped a few string queries with the gql() helper, which is now required in this version of package:graphql.

This is part of the continued quest to evaluate if this is the right substrate to build the GitHub queries I wish to or not.

PLEASE feel no pressure to merge these if they are not helpful to you.  I don't want my hacking exploration to cause disruption to existing processes.

I tested prs_landed as well as the unittests and they still ran correctly without exception.